requirements.txt
flask
tensorflow
numpy
opencv-python
librosa
Pillow
mediapipe
seaborn
